:Debuggging

+++ app/db/tables/Account.scala
@@ -1,3 +1,3 @@
 package db.tables
 
-case class Account(id:Int,accName:String,accNumber:String,accBalance:Option[Float])
+case class Account(id:Long,accName:String,accNumber:String,accBalance:Option[Float])
